refactor(api): PR-6.5 commit 3 — download/upload/preprocess 14 routes
server.py 1795 → 1354 行（−441），projects 域第三刀（数据获取 + 预处理）。

  api/routers/projects/ingestion.py  14 routes  download estimate/start/status,
                                                upload/upload-from-path,
                                                preprocess start/status/files/
                                                duplicates-removed/crop workspace/
                                                crop start/files reset+restore/thumb
  api/schemas/ingestion.py            6 BaseModel  DownloadRequest /
                                                   EstimateRequest /
                                                   UploadFromPathBody /
                                                   PreprocessStartRequest /
                                                   PreprocessRestoreRequest /
                                                   CropRect / PreprocessCropRequest

  api/routers/projects/_shared.py     + _publish_job_state（commit 5 tag / reg 也用）

清理：删 `PreprocessDeleteRequest = PreprocessRestoreRequest` 死别名（"前端
切换期间" 注释已过期）。

回归：PR-1 安全网 152 + project_endpoints 49 + preprocess_endpoints + download_endpoints = **249/249 全绿**。
route snapshot 不变（160 routes）。
