+++ studio/api/app.py
@@ -36,6 +36,7 @@ from .routers import (
 )
 from .routers.projects import crud as projects_crud
 from .routers.projects import exports as projects_exports
+from .routers.projects import ingestion as projects_ingestion
 from .routers.queue import io as queue_io_router
 from .routers.queue import lifecycle as queue_lifecycle
 from .routers.queue import outputs as queue_outputs
@@ -78,3 +79,5 @@ app.include_router(projects_crud.router)
 # PR-6.5 commit 2: train.zip / bundle.zip / export-bundle / import-bundle (path/upload) /
 # import-train（6 routes）
 app.include_router(projects_exports.router)
+# PR-6.5 commit 3: download/upload + preprocess (14 routes)
+app.include_router(projects_ingestion.router)
